Move uefi_compression to library.uefi
